test: move audit policy tests to their own package
